requirements.txt
pyspark==3.5.0
pandas<2.0.0
plotly
streamlit
google-generativeai
huggingface_hub
python-dotenv
prefect>=2.19.0,<3.0.0
dbt-duckdb>=1.6.0
great_expectations==0.18.19
duckdb
pytest>=7.0.0
pylint>=3.0.0
sqlfluff>=2.0.0
sqlfluff-templater-dbt
ruff>=0.1.0
pytest-mock
numpy<1.25.0
pydantic-settings
python-json-logger